+++ R/chevron_tlg-S4class.R
@@ -142,7 +142,7 @@ methods::setValidity("chevron_simple", function(object) {
 #'
 chevron_t <- function(main = function(adam_db, ...) build_table(basic_table(), adam_db[[1]]),
                       preprocess = function(adam_db, ...) adam_db,
-                      postprocess = std_postprocess,
+                      postprocess = report_null,
                       ...) {
   res <- .chevron_t(
     main = main,
@@ -167,7 +167,7 @@ chevron_t <- function(main = function(adam_db, ...) build_table(basic_table(), a
 #'
 chevron_l <- function(main = function(adam_db, ...) data.frame(),
                       preprocess = function(adam_db, ...) adam_db,
-                      postprocess = std_postprocess,
+                      postprocess = report_null,
                       ...) {
   res <- .chevron_l(
     main = main,
@@ -196,7 +196,7 @@ chevron_l <- function(main = function(adam_db, ...) data.frame(),
 #'
 chevron_g <- function(main = function(adam_db, ...) ggplot2::ggplot(),
                       preprocess = function(adam_db, ...) adam_db,
-                      postprocess = std_postprocess,
+                      postprocess = report_null,
                       ...) {
   res <- .chevron_g(
     main = main,