+++ R/aet01.R
@@ -171,7 +171,7 @@ aet01_post <- function(tlg, prune_0 = FALSE, ...) {
   if (prune_0) {
     tlg <- smart_prune(tlg)
   }
-  report_null(tlg)
+  std_postprocessing(tlg)
 }
 
 #' `AET01` Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.