+++ R/aet01_aesi.R
@@ -213,7 +213,7 @@ aet01_aesi_post <- function(tlg, prune_0 = FALSE, ...) {
   if (prune_0) {
     tlg <- smart_prune(tlg)
   }
-  report_null(tlg)
+  std_postprocessing(tlg)
 }
 
 #' `AET01_AESI` Table 1 (Default) Adverse Event of Special Interest Summary Table.