+++ R/egt03.R
@@ -156,7 +156,7 @@ egt03_pre <- function(adam_db, ...) {
 egt03_post <- function(tlg, prune_0 = FALSE, ...) {
   if (prune_0) tlg <- smart_prune(tlg)
 
-  report_null(tlg)
+  std_postprocessing(tlg)
 }
 
 #' `EGT03` Shift Table of ECG Interval Data - Baseline versus Minimum or Maximum Post-Baseline.