+++ R/ext01.R
@@ -135,7 +135,7 @@ ext01_pre <- function(adam_db,
 #'
 ext01_post <- function(tlg, prune_0 = TRUE, ...) {
   if (prune_0) tlg <- smart_prune(tlg)
-  report_null(tlg)
+  std_postprocessing(tlg)
 }
 
 #' `EXT01` Exposure Summary Table.