+++ tests/testthat/test-default_tlg.R
@@ -1,7 +1,7 @@
 test_that("ael01_nollt function with default argument value return expected result with test data", {
   pre_data <- ael01_nollt_pre(syn_data)
   raw_res <- ael01_nollt_main(pre_data)
-  res <- report_null(raw_res)
+  res <- std_postprocessing(raw_res)
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })