+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add `AEL02`, `AEL03` templates.
 * Modify the post processing of `MHT01` to allow multiple `row_split_var`.
+* Export the `std_postprocessing` function to simplify the post processing.
 
 # chevron 0.2.6