+++ NEWS.md
@@ -1,4 +1,4 @@
-# chevron 0.2.6.9008
+# chevron 0.2.6.9009
 
 * Add `AEL02`, `AEL03` templates.
 * Modify the post processing of `MHT01` to allow multiple `row_split_var`.